// src/app/model/trworkplace.ts
export class trworkplace {
    workplaceId: number;
    idcard: string;
    placeName: string;
    address: string;
    province: string;
    amphure: string;
    district: string;
    zipCode: string;
    telephone1: string;
    telephone2: string;
    telephone3: string;
    active: string;
    createBy: string;
    createDate: Date;
    updateBy: string;
    updateDate: Date;
};